+++ NEWS.md
@@ -4,6 +4,8 @@
 
 * Removed the `"tbl_summary-arg:statistic"` theme that was incorrectly added to `tbl_continuous()`.
 
+* Added argument `tbl_ard_summary(overall)`. When `TRUE`, the ARD is parsed into primary ARD and the Overall ARD and we run `tbl_ard_summary() |> add_overall()`. (#1940)
+
 # gtsummary 2.0.2
 
 Updates to address regressions in the v2.0.0 release: